Migrating my code with TODO comments for easy review. @m-schen, please look into them later.
