+++ pkg/sql/plan/function2/builtinSet_binary_temp.go
@@ -15,308 +15,17 @@
 package function2
 
 import (
-	"bytes"
 	"fmt"
 	"github.com/matrixorigin/matrixone/pkg/common/moerr"
 	"github.com/matrixorigin/matrixone/pkg/container/types"
 	"github.com/matrixorigin/matrixone/pkg/container/vector"
-	"github.com/matrixorigin/matrixone/pkg/sql/plan/function2/function2Util"
 	"github.com/matrixorigin/matrixone/pkg/util/fault"
-	"github.com/matrixorigin/matrixone/pkg/vectorize/datediff"
 	"github.com/matrixorigin/matrixone/pkg/vectorize/floor"
-	"github.com/matrixorigin/matrixone/pkg/vectorize/instr"
-	"github.com/matrixorigin/matrixone/pkg/vectorize/timediff"
 	"github.com/matrixorigin/matrixone/pkg/vm/process"
 	"golang.org/x/exp/constraints"
 	"math"
 	"strconv"
-	"strings"
 )
-
-// STARTSWITH
-
-func StartsWith(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) (err error) {
-	p1 := vector.GenerateFunctionStrParameter(ivecs[0])
-	p2 := vector.GenerateFunctionStrParameter(ivecs[1])
-	rs := vector.MustFunctionResult[uint8](result)
-
-	//TODO: ignoring 4 switch cases: Original code:https://github.com/m-schen/matrixone/blob/0c480ca11b6302de26789f916a3e2faca7f79d47/pkg/sql/plan/function/builtin/binary/startswith.go#L36
-	for i := uint64(0); i < uint64(length); i++ {
-		v1, null1 := p1.GetStrValue(i)
-		v2, null2 := p2.GetStrValue(i)
-		if null1 || null2 {
-			if err = rs.Append(0, true); err != nil {
-				return err
-			}
-		} else {
-			res := hasPrefix(v1, v2)
-			if err = rs.Append(res, false); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-func hasPrefix(b1, b2 []byte) uint8 {
-	if len(b1) >= len(b2) && bytes.Equal(b1[:len(b2)], b2) {
-		return 1
-	}
-	return 0
-}
-
-// ENDSWITH
-
-func EndsWith(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) (err error) {
-	p1 := vector.GenerateFunctionStrParameter(ivecs[0])
-	p2 := vector.GenerateFunctionStrParameter(ivecs[1])
-	rs := vector.MustFunctionResult[uint8](result)
-
-	//TODO: ignoring 4 switch cases: Original code:https://github.com/m-schen/matrixone/blob/0c480ca11b6302de26789f916a3e2faca7f79d47/pkg/sql/plan/function/builtin/binary/endswith.go#L43
-	for i := uint64(0); i < uint64(length); i++ {
-		v1, null1 := p1.GetStrValue(i)
-		v2, null2 := p2.GetStrValue(i)
-		if null1 || null2 {
-			if err = rs.Append(0, true); err != nil {
-				return err
-			}
-		} else {
-			res := isEqualSuffix(string(v1), string(v2))
-			if err = rs.Append(res, false); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func isEqualSuffix(b1, b2 string) uint8 {
-	if len(b1) >= len(b2) && bytes.Equal([]byte(b1)[len(b1)-len(b2):], []byte(b2)) {
-		return 1
-	}
-	return 0
-}
-
-// EXTRACT
-
-func ExtractFromDatetime(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-func ExtractFromDate(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-func ExtractFromTime(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-func ExtractFromVarchar(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-
-// FINDINSET
-
-func FindInSet(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) (err error) {
-	p1 := vector.GenerateFunctionStrParameter(ivecs[0])
-	p2 := vector.GenerateFunctionStrParameter(ivecs[1])
-	rs := vector.MustFunctionResult[uint64](result)
-
-	//TODO: ignoring 4 switch cases: Original code:https://github.com/m-schen/matrixone/blob/0c480ca11b6302de26789f916a3e2faca7f79d47/pkg/sql/plan/function/builtin/binary/findinset.go#L45
-	for i := uint64(0); i < uint64(length); i++ {
-		v1, null1 := p1.GetStrValue(i)
-		v2, null2 := p2.GetStrValue(i)
-		if null1 || null2 {
-			if err = rs.Append(0, true); err != nil {
-				return err
-			}
-		} else {
-			res := findInStrList(string(v1), string(v2))
-			if err = rs.Append(res, false); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func findInStrList(str, strlist string) uint64 {
-	for j, s := range strings.Split(strlist, ",") {
-		if s == str {
-			return uint64(j + 1)
-		}
-	}
-	return 0
-}
-
-// INSTR
-
-func Instr(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) (err error) {
-	p1 := vector.GenerateFunctionStrParameter(ivecs[0])
-	p2 := vector.GenerateFunctionStrParameter(ivecs[1])
-	rs := vector.MustFunctionResult[int64](result)
-
-	//TODO: ignoring maxLen: Original code:https://github.com/m-schen/matrixone/blob/0c480ca11b6302de26789f916a3e2faca7f79d47/pkg/sql/plan/function/builtin/binary/instr.go#L32
-	for i := uint64(0); i < uint64(length); i++ {
-		v1, null1 := p1.GetStrValue(i)
-		v2, null2 := p2.GetStrValue(i)
-		if null1 || null2 {
-			if err = rs.Append(0, true); err != nil {
-				return err
-			}
-		} else {
-			var res int64
-
-			str1 := function2Util.QuickBytesToStr(v1)
-			str2 := function2Util.QuickBytesToStr(v2)
-
-			s1GoOn, s2GoOn := len(str1) > 1, len(str2) > 1
-			//TODO Validate: Original code: https://github.com/m-schen/matrixone/blob/0c480ca11b6302de26789f916a3e2faca7f79d47/pkg/vectorize/instr/instr.go#L74
-			if s1GoOn && s2GoOn {
-				res = instr.Single(str1, str2)
-			} else if s1GoOn {
-				res = instr.Single(str1, str2)
-			} else {
-				res = instr.Single(str2, str1)
-			}
-
-			if err = rs.Append(res, false); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// LEFT
-
-func Left(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) (err error) {
-	p1 := vector.GenerateFunctionStrParameter(ivecs[0])
-	p2 := vector.GenerateFunctionFixedTypeParameter[int64](ivecs[1])
-	rs := vector.MustFunctionResult[types.Varlena](result)
-
-	for i := uint64(0); i < uint64(length); i++ {
-		v1, null1 := p1.GetStrValue(i)
-		v2, null2 := p2.GetValue(i)
-		if null1 || null2 {
-			if err = rs.AppendBytes(nil, true); err != nil {
-				return err
-			}
-		} else {
-			//TODO: Ignoring 4 switch cases: https://github.com/m-schen/matrixone/blob/0c480ca11b6302de26789f916a3e2faca7f79d47/pkg/sql/plan/function/builtin/binary/left.go#L38
-			res := evalLeft(function2Util.QuickBytesToStr(v1), v2)
-			if err = rs.AppendBytes(function2Util.QuickStrToBytes(res), false); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func evalLeft(str string, length int64) string {
-	runeStr := []rune(str)
-	leftLength := int(length)
-	if strLength := len(runeStr); leftLength > strLength {
-		leftLength = strLength
-	} else if leftLength < 0 {
-		leftLength = 0
-	}
-	return string(runeStr[:leftLength])
-}
-
-//POW
-
-func Power(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) (err error) {
-	p1 := vector.GenerateFunctionFixedTypeParameter[float64](ivecs[0])
-	p2 := vector.GenerateFunctionFixedTypeParameter[float64](ivecs[1])
-	rs := vector.MustFunctionResult[float64](result)
-
-	for i := uint64(0); i < uint64(length); i++ {
-		v1, null1 := p1.GetValue(i)
-		v2, null2 := p2.GetValue(i)
-		if null1 || null2 {
-			if err = rs.Append(0, true); err != nil {
-				return err
-			}
-		} else {
-			//TODO: Ignoring 4 switch cases:https://github.com/m-schen/matrixone/blob/0c480ca11b6302de26789f916a3e2faca7f79d47/pkg/sql/plan/function/builtin/binary/power.go#L36
-			res := math.Pow(v1, v2)
-			if err = rs.Append(res, false); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// TIMEDIFF
-
-func TimeDiff[T timediff.DiffT](ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) (err error) {
-
-	p1 := vector.GenerateFunctionFixedTypeParameter[T](ivecs[0])
-	p2 := vector.GenerateFunctionFixedTypeParameter[T](ivecs[1])
-	rs := vector.MustFunctionResult[types.Time](result)
-
-	//TODO: ignoring scale: Original code: https://github.com/m-schen/matrixone/blob/a4b3a641c3daaa10972f17db091c0eb88554c5c2/pkg/sql/plan/function/builtin/binary/timediff.go#L33
-	for i := uint64(0); i < uint64(length); i++ {
-		v1, null1 := p1.GetValue(i)
-		v2, null2 := p2.GetValue(i)
-		if null1 || null2 {
-			if err = rs.Append(0, true); err != nil {
-				return err
-			}
-		} else {
-			res, _ := timeDiff(v1, v2)
-			if err = rs.Append(res, false); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func timeDiff[T timediff.DiffT](v1, v2 T) (types.Time, error) {
-	tmpTime := int64(v1 - v2)
-	// different sign need to check overflow
-	if (int64(v1)>>63)^(int64(v2)>>63) != 0 {
-		if (tmpTime>>63)^(int64(v1)>>63) != 0 {
-			// overflow
-			isNeg := int64(v1) < 0
-			return types.TimeFromClock(isNeg, types.MaxHourInTime, 59, 59, 0), nil
-		}
-	}
-
-	// same sign don't need to check overflow
-	time := types.Time(tmpTime)
-	hour, _, _, _, isNeg := time.ClockFormat()
-	if !types.ValidTime(uint64(hour), 0, 0) {
-		return types.TimeFromClock(isNeg, types.MaxHourInTime, 59, 59, 0), nil
-	}
-	return time, nil
-}
-
-// TIMESTAMPDIFF
-
-func TimestampDiff(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) (err error) {
-	p1 := vector.GenerateFunctionStrParameter(ivecs[0])
-	p2 := vector.GenerateFunctionFixedTypeParameter[types.Datetime](ivecs[1])
-	p3 := vector.GenerateFunctionFixedTypeParameter[types.Datetime](ivecs[2])
-	rs := vector.MustFunctionResult[int64](result)
-
-	//TODO: ignoring maxLen: Original code:https://github.com/m-schen/matrixone/blob/d2921c8ea5ecd9f38ad224159d3c62543894e807/pkg/sql/plan/function/builtin/multi/timestampdiff.go#L35
-	for i := uint64(0); i < uint64(length); i++ {
-		v1, null1 := p1.GetStrValue(i)
-		v2, null2 := p2.GetValue(i)
-		v3, null3 := p3.GetValue(i)
-		if null1 || null2 || null3 {
-			if err = rs.Append(0, true); err != nil {
-				return err
-			}
-		} else {
-			res, _ := datediff.TimeStampDiff(function2Util.QuickBytesToStr(v1), v2, v3)
-			if err = rs.Append(res, false); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
 
 func AddFaultPoint(ivecs []*vector.Vector, result vector.FunctionResultWrapper, proc *process.Process, length int) (err error) {
 	for i := 0; i < 5; i++ {
